177 Pull Do Whats Right section out as a separate component
