+++ pages/index.js
@@ -1,5 +1,3 @@
-import { useState, useEffect } from 'react'
-
 import {
   ApplyLink,
   ArticlePreview,
@@ -13,325 +11,259 @@ import {
 import ConnectSvg from 'public/connect.svg'
 import ElevateSvg from 'public/elevate.svg'
 import ExploreSvg from 'public/explore.svg'
-import { LogoSection } from 'sections'
+import { WhatsRight, LogoSection } from 'sections'
 import styles from 'styles/Home.module.css'
-
-import useIsScrolledFromTop from '../helpers/hooks/use-is-scrolled-from-top'
 
 const { h2, h3 } = Heading.VARIANT
 
-const Home = () => {
-  const isScrolledFromTop = useIsScrolledFromTop(0, 40, true)
-
-  const [
-    hasWaitedForTypingAnimation,
-    setHasWaitedForTypingAnimation,
-  ] = useState(false)
-
-  useEffect(() => {
-    const typingAnimationTimer = setTimeout(() => {
-      setHasWaitedForTypingAnimation(true)
-    }, 10000)
-    return () => {
-      clearTimeout(typingAnimationTimer)
-    }
-  }, [])
-
-  return (
-    <Page>
-      <section className={`${styles.hero} ${styles.dark}`}>
-        <div className={styles.sectionContent}>
-          <div className={styles.heroContent}>
-            <Heading className={styles.heroHeading}>
-              Refactor your career
-            </Heading>
-            <Text sectionText className={styles.heroText}>
-              Until now, your growth was dependent on your company.
-              <span className={styles.heroTextExtra}>We changed that.</span>
-            </Text>
-            <ApplyLink />
-            <Text className={styles.heroTagline}> </Text>
-          </div>
+const Home = () => (
+  <Page>
+    <section className={`${styles.hero} ${styles.dark}`}>
+      <div className={styles.sectionContent}>
+        <div className={styles.heroContent}>
+          <Heading className={styles.heroHeading}>Refactor your career</Heading>
+          <Text sectionText className={styles.heroText}>
+            Until now, your growth was dependent on your company.
+            <span className={styles.heroTextExtra}>We changed that.</span>
+          </Text>
+          <ApplyLink />
+          <Text className={styles.heroTagline}> </Text>
         </div>
-      </section>
-      <section
-        className={`${styles.whatsRight} ${styles.dark}
-        ${
-          isScrolledFromTop && hasWaitedForTypingAnimation
-            ? styles.whatsRightAnimateOnce
-            : ''
-        }
-        `}
-      >
-        <div className={styles.whatsRightContent}>
-          <Heading variant={h2} className={styles.whatsRightHeading}>
-            We do what's
-            <span className={styles.whatsRightHeadingHighlight}>right</span>
-            for the engineer
+      </div>
+    </section>
+    <WhatsRight />
+    <section className={styles.features}>
+      <div className={styles.featuresTransition} />
+      <div className={styles.gridSection}>
+        <div className={styles.feature}>
+          <ConnectSvg />
+          <Heading variant={h2} className={styles.featureHeading}>
+            Connect
           </Heading>
-          <div className={styles.whatsRightItems}>
-            <div className={styles.whatsRightItem}>
-              <Heading variant={h3} className={styles.whatsRightItemHeading}>
-                Stable
-              </Heading>
-              <Text className={styles.whatsRightItemText}>
-                We hire you while you try out different startups
-              </Text>
-            </div>
-            <div className={styles.whatsRightItem}>
-              <Heading variant={h3} className={styles.whatsRightItemHeading}>
-                Streamlined
-              </Heading>
-              <Text className={styles.whatsRightItemText}>
-                Our team vets high-potential startups for you
-              </Text>
-            </div>
-            <div className={styles.whatsRightItem}>
-              <Heading variant={h3} className={styles.whatsRightItemHeading}>
-                Researched
-              </Heading>
-              <Text className={styles.whatsRightItemText}>
-                No more technical interviews
-              </Text>
-            </div>
-          </div>
+          <Text>
+            Build meaningful relationships in a distributed community of
+            Software Engineers
+          </Text>
         </div>
-      </section>
-      <section className={styles.features}>
-        <div className={styles.featuresTransition} />
-        <div className={styles.gridSection}>
-          <div className={styles.feature}>
-            <ConnectSvg />
-            <Heading variant={h2} className={styles.featureHeading}>
-              Connect
-            </Heading>
-            <Text>
-              Build meaningful relationships in a distributed community of
-              Software Engineers
-            </Text>
-          </div>
-          <div className={styles.feature}>
-            <ExploreSvg />
-            <Heading variant={h2} className={styles.featureHeading}>
-              Explore
-            </Heading>
-            <Text>
-              Build your network in a distributed community of Software
-              Engineers
-            </Text>
-          </div>
-          <div className={styles.feature}>
-            <ElevateSvg />
-            <Heading variant={h2} className={styles.featureHeading}>
-              Elevate
-            </Heading>
-            <Text>
-              Refine or expand your expertise through hands-on projects,
-              peer-to-peer coaching & professional development opportunities
-            </Text>
-          </div>
+        <div className={styles.feature}>
+          <ExploreSvg />
+          <Heading variant={h2} className={styles.featureHeading}>
+            Explore
+          </Heading>
+          <Text>
+            Build your network in a distributed community of Software Engineers
+          </Text>
         </div>
-      </section>
-      <section className={styles.articles}>
-        <Heading variant={h2} className={styles.articlesHeading}>
-          Featured Articles
+        <div className={styles.feature}>
+          <ElevateSvg />
+          <Heading variant={h2} className={styles.featureHeading}>
+            Elevate
+          </Heading>
+          <Text>
+            Refine or expand your expertise through hands-on projects,
+            peer-to-peer coaching & professional development opportunities
+          </Text>
+        </div>
+      </div>
+    </section>
+    <section className={styles.articles}>
+      <Heading variant={h2} className={styles.articlesHeading}>
+        Featured Articles
+      </Heading>
+      <div className={styles.gridSection}>
+        <ArticlePreview
+          imgSrc="/person-looking-at-monitor.jpeg"
+          imgAlt="Person looking at computer monitor"
+          heading="My job at Commit is to create an incredible experience for engineers"
+          text="One August in the 1990s, I drove across Canada, with a good friend.
+          I have so many fond memories of the journey: our debates about how
+          to solve the world's problems, the lengthy and..."
+          link="https://blog.commit.dev/articles/b5i3amkjq2nz7y4apq4jsfd66hudu8"
+        />
+        <ArticlePreview
+          imgSrc="/lillian-liang.jpeg"
+          imgAlt="Lillian Liang"
+          heading="Lillian Liang, Plastiq Inc."
+          text="Lillian Liang is a Commit engineer who joined Plastiq earlier this
+          year. We talked to her about her journey to Commit and Plastiq..."
+          link="https://blog.commit.dev/articles/committed-to-success-lillian-liang-plastiq-inc"
+        />
+        <ArticlePreview
+          imgSrc="/monolith-diagram.png"
+          imgAlt="Monolith architecture"
+          heading="Streaming out the monolith"
+          text="As a monolithic code base builds up over time, it can become harder
+          and harder to maintain, and more difficult to extract data from,
+          especially if..."
+          link="https://blog.commit.dev/articles/streaming-out-the-monolith"
+        />
+      </div>
+    </section>
+    <section className={`${styles.about} ${styles.grey}`}>
+      <div className={styles.flexSection}>
+        <Heading variant={h2} className={styles.aboutHeading}>
+          How it Works
         </Heading>
-        <div className={styles.gridSection}>
-          <ArticlePreview
-            imgSrc="/person-looking-at-monitor.jpeg"
-            imgAlt="Person looking at computer monitor"
-            heading="My job at Commit is to create an incredible experience for engineers"
-            text="One August in the 1990s, I drove across Canada, with a good friend.
-            I have so many fond memories of the journey: our debates about how
-            to solve the world's problems, the lengthy and..."
-            link="https://blog.commit.dev/articles/b5i3amkjq2nz7y4apq4jsfd66hudu8"
-          />
-          <ArticlePreview
-            imgSrc="/lillian-liang.jpeg"
-            imgAlt="Lillian Liang"
-            heading="Lillian Liang, Plastiq Inc."
-            text="Lillian Liang is a Commit engineer who joined Plastiq earlier this
-            year. We talked to her about her journey to Commit and Plastiq..."
-            link="https://blog.commit.dev/articles/committed-to-success-lillian-liang-plastiq-inc"
-          />
-          <ArticlePreview
-            imgSrc="/monolith-diagram.png"
-            imgAlt="Monolith architecture"
-            heading="Streaming out the monolith"
-            text="As a monolithic code base builds up over time, it can become harder
-            and harder to maintain, and more difficult to extract data from,
-            especially if..."
-            link="https://blog.commit.dev/articles/streaming-out-the-monolith"
-          />
+        <div className={styles.aboutDescription}>
+          <Text>
+            Commit provides the space for Software Engineers to build
+            relationships, experience growth at their own pace and establish a
+            deeper network. We hire remote-first Engineers who want to eliminate
+            the risk and the friction of finding and working for startups in
+            Silicon Valley.
+          </Text>
+          <Text>
+            Commit Engineers are matched with startup projects on a three month
+            basis, and if there is a fit with the startup, we facilitate the
+            transition to full time at that company. Otherwise, there is no
+            obligation to join and you can get matched with a new startup
+            instead.
+          </Text>
+          <Text>
+            While at Commit, we provide coaching opportunities, professional
+            development and internal projects to help develop our own open
+            source platforms.
+          </Text>
+          <Text>
+            Even after Commit Engineers ‘graduate’ to startups, they remain a
+            member of the Community so they can continue to tap into the
+            knowledge base and join craft, skills or interest based groups.
+          </Text>
         </div>
-      </section>
-      <section className={`${styles.about} ${styles.grey}`}>
-        <div className={styles.flexSection}>
-          <Heading variant={h2} className={styles.aboutHeading}>
-            How it Works
+      </div>
+    </section>
+    <section className={styles.platform}>
+      <div className={styles.platformContainer}>
+        <div className={styles.platformTopRow}>
+          <Heading variant={h2} className={styles.platformHeading}>
+            The Commit Platform
           </Heading>
-          <div className={styles.aboutDescription}>
-            <Text>
-              Commit provides the space for Software Engineers to build
-              relationships, experience growth at their own pace and establish a
-              deeper network. We hire remote-first Engineers who want to
-              eliminate the risk and the friction of finding and working for
-              startups in Silicon Valley.
-            </Text>
-            <Text>
-              Commit Engineers are matched with startup projects on a three
-              month basis, and if there is a fit with the startup, we facilitate
-              the transition to full time at that company. Otherwise, there is
-              no obligation to join and you can get matched with a new startup
-              instead.
-            </Text>
-            <Text>
-              While at Commit, we provide coaching opportunities, professional
-              development and internal projects to help develop our own open
-              source platforms.
-            </Text>
-            <Text>
-              Even after Commit Engineers ‘graduate’ to startups, they remain a
-              member of the Community so they can continue to tap into the
-              knowledge base and join craft, skills or interest based groups.
-            </Text>
-          </div>
+          <Text sectionText>
+            The Commit Platform replaces the technical interview with deep
+            backgrounds, matches Software Engineers with relevant startups, and
+            connect Engineers with peers for interest sharing or technical
+            support.
+          </Text>
         </div>
-      </section>
-      <section className={styles.platform}>
-        <div className={styles.platformContainer}>
-          <div className={styles.platformTopRow}>
-            <Heading variant={h2} className={styles.platformHeading}>
-              The Commit Platform
-            </Heading>
-            <Text sectionText>
-              The Commit Platform replaces the technical interview with deep
-              backgrounds, matches Software Engineers with relevant startups,
-              and connect Engineers with peers for interest sharing or technical
-              support.
-            </Text>
-          </div>
-          <div className={styles.platformFlex}>
-            <div className={styles.gridSection}>
-              <div className={styles.profileColumnLeft}>
-                <picture>
-                  <source
-                    media="(min-width:1201px)"
-                    srcSet="/platform-profile-large.png"
-                  />
-                  <source
-                    media="(min-width:1024px) and (max-width:1200px)"
-                    srcSet="/platform-profile-desktop.png"
-                  />
-                  <source
-                    media="(min-width:770px) and (max-width:1023px)"
-                    srcSet="/platform-profile-tablet.png"
-                  />
-                  <img
-                    src="/platform-profile-mobile.png"
-                    alt="Screenshot of Commit's Engineering Parnter profile page"
-                    className={styles.platformImage}
-                  />
-                </picture>
-              </div>
-              <div className={styles.profileColumnRight}>
-                <Heading variant={h3} className={styles.platformTitle}>
-                  Platform profile
-                </Heading>
-                <Text sectionText className={styles.platformDescription}>
-                  Get the latest community events, news and updates. Most
-                  importantly, begin to look for startup opportunities
-                  recommended specifically for you. This is also where engineers
-                  begin to connect with their peers and join interest-based
-                  groups.
-                </Text>
-              </div>
+        <div className={styles.platformFlex}>
+          <div className={styles.gridSection}>
+            <div className={styles.profileColumnLeft}>
+              <picture>
+                <source
+                  media="(min-width:1201px)"
+                  srcSet="/platform-profile-large.png"
+                />
+                <source
+                  media="(min-width:1024px) and (max-width:1200px)"
+                  srcSet="/platform-profile-desktop.png"
+                />
+                <source
+                  media="(min-width:770px) and (max-width:1023px)"
+                  srcSet="/platform-profile-tablet.png"
+                />
+                <img
+                  src="/platform-profile-mobile.png"
+                  alt="Screenshot of Commit's Engineering Parnter profile page"
+                  className={styles.platformImage}
+                />
+              </picture>
             </div>
-            <div className={`${styles.gridSection} ${styles.opportunitiesRow}`}>
-              <div className={styles.opportunitiesColumnLeft}>
-                <picture>
-                  <source
-                    media="(min-width:1201px)"
-                    srcSet="/platform-opportunities-large.png"
-                  />
-                  <source
-                    media="(min-width:1024px) and (max-width:1200px)"
-                    srcSet="/platform-opportunities-desktop.png"
-                  />
-                  <source
-                    media="(min-width:770px) and (max-width:1023px)"
-                    srcSet="/platform-opportunities-tablet.png"
-                  />
-                  <img
-                    src="/platform-opportunities-mobile.png"
-                    alt="Screenshot of Commit's recommended opportunities page"
-                    className={styles.platformImage}
-                  />
-                </picture>
-              </div>
-              <div className={styles.opportunitiesColumnRight}>
-                <Heading variant={h3} className={styles.platformTitle}>
-                  Opportunities
-                </Heading>
-                <Text sectionText className={styles.platformDescription}>
-                  Commit engineers don't apply to startups. Instead, startups
-                  apply to our engineers.
-                  <br />
-                  <br />
-                  Our team works tirelessly to match engineers with the top
-                  opportunities based on their skill set, interests, and career
-                  goals. Engineers can browse new recommendations one by one,
-                  and dig into each individual opportunities, with detailed and
-                  insightful information we've gathered.
-                </Text>
-              </div>
+            <div className={styles.profileColumnRight}>
+              <Heading variant={h3} className={styles.platformTitle}>
+                Platform profile
+              </Heading>
+              <Text sectionText className={styles.platformDescription}>
+                Get the latest community events, news and updates. Most
+                importantly, begin to look for startup opportunities recommended
+                specifically for you. This is also where engineers begin to
+                connect with their peers and join interest-based groups.
+              </Text>
+            </div>
+          </div>
+          <div className={`${styles.gridSection} ${styles.opportunitiesRow}`}>
+            <div className={styles.opportunitiesColumnLeft}>
+              <picture>
+                <source
+                  media="(min-width:1201px)"
+                  srcSet="/platform-opportunities-large.png"
+                />
+                <source
+                  media="(min-width:1024px) and (max-width:1200px)"
+                  srcSet="/platform-opportunities-desktop.png"
+                />
+                <source
+                  media="(min-width:770px) and (max-width:1023px)"
+                  srcSet="/platform-opportunities-tablet.png"
+                />
+                <img
+                  src="/platform-opportunities-mobile.png"
+                  alt="Screenshot of Commit's recommended opportunities page"
+                  className={styles.platformImage}
+                />
+              </picture>
+            </div>
+            <div className={styles.opportunitiesColumnRight}>
+              <Heading variant={h3} className={styles.platformTitle}>
+                Opportunities
+              </Heading>
+              <Text sectionText className={styles.platformDescription}>
+                Commit engineers don't apply to startups. Instead, startups
+                apply to our engineers.
+                <br />
+                <br />
+                Our team works tirelessly to match engineers with the top
+                opportunities based on their skill set, interests, and career
+                goals. Engineers can browse new recommendations one by one, and
+                dig into each individual opportunities, with detailed and
+                insightful information we've gathered.
+              </Text>
             </div>
           </div>
         </div>
-      </section>
-      <section className={styles.testimonials}>
-        <div className={styles.flexSection}>
-          <Heading variant={h2} className={styles.testimonialsHeading}>
-            What Our Engineering Partners Have to Say
-          </Heading>
-          <Testimonials className={styles.testimonialsCarousel} />
-        </div>
-      </section>
-      <LogoSection title="Who We Work With">
-        <CompanyLogo
-          url="https://www.procurify.com/"
-          alt="Procurify logo"
-          logoSrc="/procurify.png"
-        />
-        <CompanyLogo
-          url="https://www.vidyard.com/"
-          alt="Vidyard logo"
-          logoSrc="/vidyard.png"
-        />
-        <CompanyLogo
-          url="https://www.planworth.co/"
-          alt="Planworth logo"
-          logoSrc="/planworth.png"
-        />
-        <CompanyLogo
-          url="https://www.dapperlabs.com/"
-          alt="Dapper logo"
-          logoSrc="/dapper.png"
-        />
-        <CompanyLogo
-          url="https://www.usepatch.com/"
-          alt="Patch logo"
-          logoSrc="/patch.png"
-        />
-        <CompanyLogo
-          url="https://www.plastiq.com/"
-          alt="Plastiq logo"
-          logoSrc="/plastiq.png"
-        />
-      </LogoSection>
-      <Footer />
-    </Page>
-  )
-}
+      </div>
+    </section>
+    <section className={styles.testimonials}>
+      <div className={styles.flexSection}>
+        <Heading variant={h2} className={styles.testimonialsHeading}>
+          What Our Engineering Partners Have to Say
+        </Heading>
+        <Testimonials className={styles.testimonialsCarousel} />
+      </div>
+    </section>
+    <LogoSection title="Who We Work With">
+      <CompanyLogo
+        url="https://www.procurify.com/"
+        alt="Procurify logo"
+        logoSrc="/procurify.png"
+      />
+      <CompanyLogo
+        url="https://www.vidyard.com/"
+        alt="Vidyard logo"
+        logoSrc="/vidyard.png"
+      />
+      <CompanyLogo
+        url="https://www.planworth.co/"
+        alt="Planworth logo"
+        logoSrc="/planworth.png"
+      />
+      <CompanyLogo
+        url="https://www.dapperlabs.com/"
+        alt="Dapper logo"
+        logoSrc="/dapper.png"
+      />
+      <CompanyLogo
+        url="https://www.usepatch.com/"
+        alt="Patch logo"
+        logoSrc="/patch.png"
+      />
+      <CompanyLogo
+        url="https://www.plastiq.com/"
+        alt="Plastiq logo"
+        logoSrc="/plastiq.png"
+      />
+    </LogoSection>
+    <Footer />
+  </Page>
+)
 
 export default Home
