+++ sections/index.js
@@ -2,3 +2,4 @@ export { default as DiversityAndInclusionSection } from './diversity-and-inclusi
 export { default as ExecutiveTeamSection } from './executive-team-section'
 export { default as OpensourceSection } from './opensource-section'
 export { default as LogoSection } from './logo-section'
+export { default as WhatsRight } from './do-whats-right-section'
